Refactor: python, cts: Move StartOnebyOne into the _cts module.
No code changes.
