+++ cts/lab/CTStests.py
@@ -33,40 +33,6 @@ AllTestClasses = [ ]
 AllTestClasses.append(FlipTest)
 AllTestClasses.append(RestartTest)
 AllTestClasses.append(StonithdTest)
-
-
-class StartOnebyOne(CTSTest):
-    '''Start all the nodes ~ one by one'''
-    def __init__(self, cm):
-        CTSTest.__init__(self,cm)
-        self.name = "StartOnebyOne"
-        self.stopall = SimulStopLite(cm)
-        self._start = StartTest(cm)
-        self.ns = NodeStatus(cm.Env)
-
-    def __call__(self, dummy):
-        '''Perform the 'StartOnebyOne' test. '''
-        self.incr("calls")
-
-        #        We ignore the "node" parameter...
-
-        #        Shut down all the nodes...
-        ret = self.stopall(None)
-        if not ret:
-            return self.failure("Test setup failed")
-
-        failed = []
-        self.set_timer()
-        for node in self._env["nodes"]:
-            if not self._start(node):
-                failed.append(node)
-
-        if len(failed) > 0:
-            return self.failure("Some node failed to start: " + repr(failed))
-
-        return self.success()
-
-#        Register StartOnebyOne as a good test to run
 AllTestClasses.append(StartOnebyOne)
